Add (temporary) capability Tmg Geolocation
The official "Geolocation" capability has NO presentation currently. To
be able to display at least "Speed", we need to have a temporary
version of the capability, so that we can display that attribute.

We should continue to publish both the official and temp attributes,
until we're sure we can remove the temp one.